Appear blocks use a dict instead and don't need the process function
Also added a return function if block is reinited so it doesnt still spawn more, before this was spawning over and over again.

Just uses the any_side function for reiniting and activating so blocks properly disappear and reappear at the right time, two ground blocks are not active at once now.