Silence get_node() null error
`get_node()` emits an error to the debugger when it can't find anything and returns null.

`get_node_or_null()` does not emit an error when it can't find anything.

It seems intentional that the `existing_bump_effect` variable can be null, so `get_node_or_null()` should be used here